feat: Translate from American to British
The core functionality is ready.
Replacing was fairly easy. But there is the problem of cases.
The replacing lists are all in lowercase, but you have to output the same case as the input.
I implemented a solution for the titles, which works.
What you do is use case insensitive regex to match without manipulating anything.
Then you replace. And when you replace, you just Uppercase the first character.